# docs/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Studio",
    "version": "1.0.3",
    "description": "Agent-facing API. Branch on the `code` field of an error, never on its prose.\nIn server mode every call carries `Authorization: Bearer avs_…`; scopes are read, produce, publish, admin.\nName the channel explicitly (body.channelId, ?channel= or X-AVS-Channel) — the active channel is only a default.\nSend an Idempotency-Key (or clientRef on a create) so a retry after a timeout cannot spend twice.\nPoll GET /events?after=<id> for progress; a WebSocket at /ws carries the same feed live."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "error"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine word. Branch on this, never on `error`."
          },
          "error": {
            "type": "string",
            "description": "Short reason, in the user's language unless ?lang= says otherwise."
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Mint with `npm run token -- create` on the machine running the server."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/topics/suggest": {
      "post": {
        "summary": "Channel-voiced topic proposals, deduped against everything already made or scheduled. Data only — nothing starts.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Channel id. Also accepted as X-AVS-Channel or body.channelId. Defaults to the token’s channel, then the active one."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/topics/history": {
      "get": {
        "summary": "Past proposals and what became of them.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "read",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Channel id. Also accepted as X-AVS-Channel or body.channelId. Defaults to the token’s channel, then the active one."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/topics/{id}/accept": {
      "post": {
        "summary": "Turn one proposal into a project and start it. This is the only path from a suggestion to a paid run.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: suggestion_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/topics/{id}/schedule": {
      "post": {
        "summary": "Put a proposal on the calendar. Creates a slot, not a job.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Known codes: suggestion_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/calendar": {
      "get": {
        "summary": "Scheduled slots.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/plan": {
      "post": {
        "summary": "Fill the coming days from pending proposals. Slots only.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Channel id. Also accepted as X-AVS-Channel or body.channelId. Defaults to the token’s channel, then the active one."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/{id}": {
      "delete": {
        "summary": "Cancel a slot; its idea returns to the pool.",
        "tags": [
          "assistant"
        ],
        "x-avs-scope": "produce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "token_required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "scope_denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}